catalyst/runners refactoring (solved cyclic imports)
